fix: Remove GPU code, rely on a running drawcallback
This changes the plugin from relying on its own implementation of DrawCallbacks, to instead wrapping around an active DrawCallbacks.

This means it shouldn't conflict any more with DrawCallbacks plugins, and lets it work in GPU mode with things like 117HD.

This means you need one of currently GPU or 117HD running for the GPU mode to work though.